Microsoft-Graph-RAG/requirements.txt
flask
flask_cors
PyMuPDF
python-dotenv
pillow
requests
pyyaml
graphrag
